* ext/socket/socket.c (socket_s_ip_address_list): Cast EXTRA_SPACE as   int.  This suppress a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@40597 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/socket/socket.c
@@ -1759,7 +1759,7 @@ socket_s_ip_address_list(VALUE self)
 #elif defined(SIOCGIFCONF)
     int fd = -1;
     int ret;
-#define EXTRA_SPACE (sizeof(struct ifconf) + sizeof(union_sockaddr))
+#define EXTRA_SPACE ((int)(sizeof(struct ifconf) + sizeof(union_sockaddr)))
     char initbuf[4096+EXTRA_SPACE];
     char *buf = initbuf;
     int bufsize;
